(PUP-4503,PUP-4540,PUP-2023) Update service provider default for newer debian/ubuntu releases
With the release of Debian 8 Jessie and Ubuntu 15.04 (Vivid Vervet),
ubuntu and debian are noew defaulting the available service manager to
systemd, rather than init or upstart. This commit updates Puppet to
use systemd as the default service provider on these two systems, while
also defaulting back to other service providers if necessary.