(PUP-4436) Generate error message in eventlog differently
Prior to commit 88ae8a9, it was an error to specify a `confdir` whose
parent directory didn't exist. After the change, puppet would create the
ancestor directories `C:/does/not`, enabling puppet's settings catalog
to create the `confdir` named `C:/does/not/exist`.

This commit changes the test to no longer rely on setting an invalid
confdir.  Instead it just runs `puppet agent` trying to connect to a
non-existent master, as it did before, and it changes the expectation to
match the error we generate.

The commit also changes from the deprecated `puppet_agent` beaker
method to `puppet('agent', ...)`.

It would have been nice to test this behavior using:

    puppet apply -e "err('some message')"

However, `apply` uses the console as the default log destination. It is
possible to specify `puppet apply --logdest eventlog ...` but that
doesn't test the default `agent` behavior of always writing to the
eventlog.

[skip ci]