(PUP-4619) parsedfile: add a self.order hook that defaults to noop
It should be possible for parsedfile providers to enforce an order
on the records that are written to disk. This is necessary because
resource ordering is only effective within the same transaction.
Parsedfile will often add new records in subsequent transactions,
whenever new resources appear in the node's catalog.

Unless overridden, self.order is an identity function and does not
alter the order at all.