(PUP-1295) Fix always purging package resources that are not installed.
For package providers, the base package provider was defaulting to a
property hash with a status of :absent when the package was missing.

When the resource is `ensure => purged`, an :absent status causes it to
be purged.  This is because an absent package could represent a
"partially installed" package that is not truly installed and can
therefore be purged (see the dpkg provider).  If the base provider
always returns a status of :absent for missing packages then purgeable
providers will always attempt to purge a missing resource.

The fix is to have the base provider default to a :purged status when
the resource is missing and the provider supports purging.  Otherwise,
it defaults to :absent like before.