(PUP-2573) Make agent locker raise LockError if we fail to acquire lock
This fixes a possible raise condition in the puppet agent where it first
checked if the lock was taken and then later on aquired it, possibly
having some other agent taking the lock in between with no handling of
that situation.

Also it updates the description of the lock method to actually match
what it does, namely return the value of the block itself. Not
true/false based on if lock was acquired. That was in fact impossible to
find out previously.