(maint) Update service to work on debian
Unlike redhat, debian stores env files in '/etc/default'. We need to be
able to search there in order to access them. This commit adds that
search path to the systemd service file. Note that this works because
the service doesn't depend on the existance of those files. So, on
redhat, neither '/etc/default/puppet' nor '/etc/default/puppetagent'
will exist, but the service will continue to run regardless.