(PUP-2630) Populate $server_facts during puppet apply
Prior to this commit, $server_facts would only be populated during
an agent run. This it made it difficult for users to use
`$server_facts[environment]` in their puppet code since they could
not test it via a puppet apply.

Now that variable will be populated during an apply if
trusted_server_facts is true. It will only contain the environment
portion of the hash.