(PUP-4538) Preserve permissions when downloading external facts
Previously, `puppet plugin download` did not preserve permissions when
downloading external facts. As a result, executable external facts
would not be executable after download, and their fact values not
resolved.

This was caused because the `plugin` face duplicated the logic for
constructing `Puppet::Configurer::Downloader` objects and didn't
specify `source_permissions => use` when downloading external facts.

This commit changes the `download` action to reuse the same logic that
the `Puppet::Configurer` uses to download plugins and external facts.
It is not a pure refactor, because the `Puppet::Configurer::PluginHandler`
invokes `Puppet::Util::Autoload.reload_changed`, something that the
old implementation did not do. However, since `puppet plugin download`
is a command line app that exits, the reloading should not cause any
bad side effects.

This commit also adds tests for the `plugin` face, as the previous tests
were useless.