(PUP-4458) Refactor validation of parameter signatures
This commit unifies how parameters are validated so that instantiations
of nodes, classes, and definitions use a validatio logic that has more
in common with the validation of lambda invocations and function calls.

This commit also adds deprecation warnings to internal methods that
are no longer used by Puppet but kept for backward compatibility.