(PUP-4619) add bind mount support to fstab ordering
Bind mounts on Linux and other platforms allow mounting an existing
diretory at another file system location, e.g.

/opt/data02/mysql /var/lib/mysql none bind,rw 0 0

When the source is moved to a new filesystem that gets its own mount entry,
it must appear before the bind mount, so that the latter will be available
at boot time. Just appending it would not work.

/opt/data02/mysql /var/lib/mysql bind,rw 0 0
/dev/vg0/data2 /opt/data02 ext4 defaults,ro 0 2

Since /opt/data02/mysql is now part of a separate filesystem, it is not
available before /opt/data02 is mounted, and consequently cannot be
bind-mounted. The fstab entries must be reversed.