(maint) Archive spec_order.txt on appveyor failures
Previously, tests executed in a platform-specific order in appveyor, and
if an order dependent failure happened, we didn't know what the order
was.

This commit sets the `LOG_SPEC_ORDER` environment variable which causes
puppet's spec helper to write each spec to `spec_order.txt` in the
project root.

It also adds an `on_failure` handler to archive the file, which can then
be downloaded from the job's Artifacts tab.