(PUP-3742) Smoke test gems
Creates a rake task `ci:test:gem` and an acceptance test for installing
puppet as a gem and smoke testing the install against a predefined set
of hosts windows (ruby x86 and x64), osx and redhat. The osx case is
needed because we have a darwin-specific facter gem that has a
dependency on CFPropertyList.

The `ci:test:gem` task relies on the existing git setup step that
installs `git` and `ruby` on each host, even though we really only need
`ruby`.

Also on Windows, we're using the our vendored ruby, which already
includes some gems, e.g. win32-service, ffi. So the test setup step
wipes the current set of installed gems. Alternatively, it would be nice
to only install our vendored ruby, but we do not current publish that as
a build artifact.

[skip ci]