(PUP-4131) Implement insync? for gem package provider.
The gem package provider was previously not implementing insync? but
supported a provider-specific version string for gem version specifiers
(e.g. '~> 2.0').  This resulted in the resource always being out-of-sync
when a gem version specifier was used.

This adds an implementation of insync? that uses the rubygems API to
verify the version specifier matches.