(PUP-4652) Don't rely on beaker's `agent` method
Previously, the test passed when run with a single agent node, but
failed when run with multiple agent nodes, with the error:

    TypeError Exception: no implicit conversion of String into Integer

The root cause is that Beaker metaprograms methods for each role,
where the method returns either a single host or an array of hosts
with that role. And in the latter case, we were calling `['platform']`
on an array causing the TypeError.

This commit modifies the test to not assume there is a single agent.
Instead we always create unix and windows external facts on the master,
and for each agent, verify the external fact is resolved.

[skip ci]