(maint) Implement inspect method on base type provider.
It looks like RSpec was changed to call inspect rather than to_s when
displaying error messages on expected values.  This results in
uninformative error messages when a type provider is passed to expect,
such as in:

```
expect(provider).to be_insync(is)
```

This simply adds an `#inspect` method that calls the existing `#to_s`
implementation to restore the expected error messages.