(PUP-2638) Use settings during puppet apply setup
Previously, if you installed puppet, and never executed `puppet agent`,
but immediately executed `puppet apply` (such as with vagrant
workflows), and `graph` was true (as is the default in PE), then
puppet would fail to write the graph. This occurred because `puppet
apply` tries to write the graph (during `Catalog#to_ral`) before
creating the necessary settings related directories (as a side-effect
of calling `Puppet::Configurer.new`).

The `agent` and `device` applications also use the Configurer to apply
the catalog, but they explicitly use the `main`, `ssl`, and `agent`
sections during application setup, and before calling `Catalog#to_ral`.

This commit modifies `apply` to match the other two. Alternatively, the
setup logic could be added to the base `Puppet::Application#setup`, but
it's more complex due to run_mode.

This commit adds an acceptance test, which uses a temporary vardir to
ensure `statedir` and `graphdir` do not exist. It also adds a debug
message to print which settings-related sections are applied.