(PUP-4668) Fix leading 'class...' 3.x resource type conversion
Before this the use of references such as Classroom, Classroom[x] would
when converted to a 3.x parameter value i.e. used this way:

notify { bad: require => Classroom[nbr3] }

would misinterpret the leading 'Class' (or 'class') as if it was a
reference to the class nbr3 not the resource Classroom[nbr3].

This was caused by a non end-anchored regular expression.