(PUP-4673) Use the json api in the pip provider
Use the json api to search for the latest available package. This was
requested by the team that runs pypi.python.org. The json response can
be cached in fastly, all XML requests have to go back to the origin.
It also means that open-uri can handle inspection of local proxies, etc.
And it means less XML is used in the world.