(PUP-3774) Add helpers to the list property type for doing set differences
The list type `should` method used to only return a string of the
joined members, based on the membership flag. For PUP-3774 I need to
actually treat it as a set that I can diff against the values returned
by the provider. This adds the necessary helpers to do that.