(PUP-3774) Use gpasswd for modifying groups in useradd provider
Previously, the useradd provider would set all of a user's groups at
once using usermod. This had a couple of problems when a user came
from a non-local PAM source

1) The user could potentially not exist in /etc/passwd. This would
simply cause `usermod` to fail.

2) The user could be in a group that's not considered valid by the
`usermod` command (especially a group that contains a space in its
name).

The new implementation uses `gpasswd` to modify the user/group
mappings only as needed. This means Puppet will completely ignore any
existing groups when adding a user to a new group, and will be able to
modify local group mappings even if the user comes from a different
PAM source.

Validation will still fail if Puppet is asked to make changes to a
user/group mapping when the group has spaces in it, as that is not a
valid operation with gpasswd (no local group would have such a name)