(PUP-3774) Only apply group changes in useradd provider
Previously, the useradd provider would set all of a user's groups at
once using usermod. This would cause problems when a user had existing
groups coming from AD, which would not be valid as arguments to
usermod.

The new implementation uses `gpasswd` to modify the user/group
mappings only as needed. This means Puppet will completely ignore any
existing groups when adding a user to a new group.

Validation will still fail if Puppet is asked to make changes to a
user/group mapping when the group has spaces in it, as that is not a
valid operation.