(PUP-4648) Fix EPP trimming/indentation problems
Before this, a right trim would replace the whitespace on each following
line up to the next tag instead of just trimming the space on the same
line as the ending tag.

A similar problem occured for a left trim, as it removed the line ending
of the previous line.

Fixing these issues means more strictness wrt which whitespace is
trimmed by each rule.
* A right trim rule trims at most to (including) a line break, stopping
on the first non whitespace.
* A left trim rule troms at most to the beginning of the same line as
the opening (left trimming tag).

This has consequences for comment tags (as they cannot start with a left
trim). Instead, they are now always made left trimming. This is the most
reasonable thing to do since they otherwise needs to be placed
immediately adjacent to a tag or text on a line, cannot be indented etc.