(PUP-4521) Add configured_environment to Node Facts find
Prior to this commit, there was no way to write an external node
classifier that allows some agents - based on Facts - to override the
node classifier with its own requested environment.

The first request from the agent for the environment will not use the
correct Facts (that might need to include custom facts), which may end
up classifying it in the incorrect environment. Subsequent requests
will send that environment instead of the configured environment.

Fix by passing the configured environment to the Node::Facts find
request as a new option.