(maint) Restore check for non-existent path
In the course of auditing differences between the puppet and pe-puppet
code bases, I found this one diff. The history on this code change is
torturous and tortuous featuring the rarely seen quadruple-revert
maneuver.

All that aside, it seems like a legit bit of defensive programming,
with no real downside, so this commit restores it.