(PUP-4709) Fix problem with optimized version of title_key_for_ref
Before this, the optimized version of title_key_for_ref was not
compatible with how brackets were handled before the optimization. This
caused titles that include a right bracket to signal the end of the type
part - e.g. "Notify[[foo]bar]" would split into type "Notify", and title
"[foo".

The fix is to search for the right bracket using rindex instead of
index.

Note that all implementations (original, optimized, and now the fixed
variant) are very brittle and are based on assumptions about the string
format of a resource "identifier". The real solution is to not use a
string representation at all. Luckily this is an internal method that
is/should not be used in user code in general.