(packaging) (re-4770) Update init script to ignore old/incorrect sysconfig values
Currently, the EL and SLES init scripts will read variables from /etc/sysconfig/puppet
and override variables set in puppet.conf or the init sctipt. This can lead to
undesired behavior, like an old sysconfig file setting variables incorrectly, causing
startup failures, etc.

This updates the sysconfig files to be more conistent with the defaults file we
ship for Debian, and updates the EL and SUSE init scripts to only use the
PUPPET_EXTRA_OPTS variable if it exists, and to ignore other settings that
are better set in puppet.conf or that should not need to be overridden at
all (like pidfile for example).