(PUP-4703) Optimize copy_as_resource checks for future parser
Before this, there would be one call to Puppet.future_parser? for every
attribute of a resource. This call is expensive esp. when it is not
given an environment as an argument. Users report 20-30% of catalog
compilation time spent on this alone!

Now, the check is only performed once per resource, and then with the
environment of the resource. This is safe because all attributes of a
resource are natually also in the same environment, and the parser
setting cannot change (at least in a sane way) between attributes being
copied.