(PUP-4752) Optimize & Secure safe_posix_fork
Before this, on POSIX systems, each time we execute a command, it forks a new process and this one will close inherited file descriptors by looping between 3 and 256 although there is more than 256 possible file descriptors. It's not safe and it's really slow.

Now, we try to list open file descriptors by reading "/proc/self/fd" and close the ones we don't want. This will fallback to the original method if it can't find "/proc/self/fd".