(PUP-4747) Change PopsBridge::Expression.to_s to return the source
Before this, the PopsBridge::Expression would return a string from to_s
that was produed by the ModelTreeDumper. That format is S-expression
like and intended for AST tree debugging.

This commit changes the return to instead be the source code associated
with the expression.

There are naturally issues with this approach as 4.x accepts potentially
complex expressions as a default value expression. The receiving end of
the result must be prepared to deal with multiple lines being returned,
and text that possibly includes comments, conditional logic, etc.

The solution should be as ok as the 3.x implementation that would have
similar issues (although not supporting complex expressions).