(PUP-4753) Add ability to call 3.x and 4.x function from scope
Before this it was close to impossible to correctly call a function in
an agnostic way from a 3.x function or ERB template.

This commit adds the ablity to call a function by given name, with
arguments given using 4.x semantics. It will load and call either a 4.x
function or a 3.x function. For a 3.x function it will perform argument
transformation as needed (just as if the call came from 4.x puppet
code).

Since there is no (easily obtained) information abailable about which
module the call originates from a 4.x call is done from the perpective
of the environment (thus, later when there are module private functions,
those will not be callable; but for now, this distinction does not
exist).