(PUP-4776) Make 'puppet agent arg' error instead of silently running the catalog
Currently, 'puppet agent' only analyses options beginning with '--', such as '--test' and
'--disable'.  Any other words after agent are ignored.  That is different from the way that
commands like 'puppet resource' and 'puppet apply' work, mainly because those require
an argument, while 'puppet agent' does not.  This has led to, for instance, one of our new
sysadmins typing 'puppet agent disable' instead of 'puppet agent --disable' and being
terribly surprised when about 5 minutes later the server died.

This patch raises an error if someone on the CLI tries to invoke 'puppet agent' with what
has to be an invalid argument.