(PUP-4781) filemode retrieved by static_compiler should be stringified
When a file resource is declared without an explicit mode, the mode
retrieved from the metadata by the static_compiler catalog terminus is
not stringified, resulting in file resources with integer file modes in
the resultant catalog and a depreciation warning such as:

Warning: Non-string values for the file mode property are deprecated. It
must be a string, either a symbolic mode like 'o+w,a+r' or an octal
representation like '0644' or '755'.
(at /usr/lib/ruby/vendor_ruby/puppet/type/file/mode.rb:69:in `block (2
levels) in <module:Puppet>')