(maint) Fix failing test caused by databinding being in effect
The evaluator spec tests started failing when Hiera 3.0.1 was brought
in. This was unexpected as the tests does not rely on data binding being
in effect. Instead it is running with string variables to ensure that no
unknown variables are used in the tests. At the same time, tests had not
made the now required variables $environment and $trusted available.

The fix is to run these tests without databinding.