(PUP-4787) Initialize @vars of Puppet:Node with expected defaults
Before this, (introduced in PUP-4775 where the serialization
was changed to use Psych support for asymetric serialization), two of
Puppet::Node's instance varaibles (@classes, and @parameters) where
initialized to nil on deserialization. When again serializing a check
was made if these where empty? (which fails).

This was caught by a roundtripping acceptance test where a
yaml-serialized node was turned into json via serialization.

The fix was to initialize to [] and {} respectively in the
initialize_from_hash method.

This also adds a unit test for a roundtrip serialization where classes
and parameters are not included in the serialized data.