(PUP-4789) Make hiera_include function call include in calling scope
Before this, the hiera_include function would call the include function
with its closure scope instead of the calling scope. This caused the
include function to not understand in which class the include was made,
and variable scoping would be wrong - not seeing variables in node
scope.

The hiera 4.x functions already received the calling scope, but did not
pass it on to the post processing method, that in hiera_include's case
performs the call to the include function.