(PUP-4787) Add additional checks for node[name] serialization
Before this there was a possibility that code could initialize a
Puppet::Node with a nil name because the method intialize_from_hash is
public. The check for name being nil was in a class method
(self.from_data_hash).

This makes the implementation more robust as it moves the check for nil
name down to where it affects the @name.

Spec tests updated to also check for nil name.