(maint) Shorten vcr fixture filename length
Previously CI was failing on windows exceeding the maximum windows path
length of 260 characters. This occurs because our jenkins instance uses
a prefix that includes the job name and each axis in the matrix:

    C:\jenkins\workspace\platform_aio-puppet_unit-ruby_master-windows\RUBY\ruby-2.1.5.2-x86\SLAVE\unit-win2012\

This commit shortens the spec path from 159 characters to 125. When
combined with the 107 character prefix, will be 232 characters, less
than the max.