(maint) Constrain ffi to less than 1.9.9
FFI gem 1.9.9 causes ruby to segfault on windows[1]. This is causing
Windows PRs on appveyor to fail, but not in our internal CI, because we
use our rubygems mirror.

This commit is a temporary measure to allow appveyor testing to pass.
Once the upstream issue is resolved, we can remove the constraint.

[1] https://github.com/ffi/ffi/issues/440