(PUP-4810) Avoid caching environments that have 0 ttl
Before this, a cached environment with 0 ttl, would still be kept
in memory until the same environment was asked for again (which
then always results in a cache eviction).

Now, instead of making an entry in the environment cache, an
environment configured with 0 ttl is simply not entered into
the cache. It will instead be garbage collected as soon as there is
nothing binding it in memory.

This was made possible by the fact that environments are now always
bound in the puppet context for as long as it is needed. (This was not
the case historically; and then the now changed behavior would have
caused the same environment to reload).