(PUP-4810) Add a more agressive environment cache eviction policy
Before this, an optimization was added that did not add 0 ttl entries
to the cache. This was not quite correct as the external service was not
kept in sync, and the environment settings associated with the
environment were not evicted. Also, it was required for an environment
to both be marked as evicted by an external service and again requested.
This was pretty much a 0-score game.

Now, with the modified algorithm all expired environments will be
evicted when any environment is requested.

To avoid a worst case where there are many cached environments expiring
a long time in the future, the algorithm keeps a sorted set of times
where an expiration will take place. It uses this, and advice from the
external cache service to determine if environments should be evicted or
not. As environments are frequntly requested, this ensures that the list
of cached environments is not traversed and checked when there is a
guarnatee that there are no expirations.

Specs had to be changed as the new algorithm is sensitive to Time vs
numeric data types in a way the previous was not. There was also a test
specific implementation of the external cache service that was not
idempotent (it changed internal state in a boolean query operation, and
the logic now calls this query operation in an additional location).