(PUP-4405) Don't override group setting
Previously, the tests were setting puppet's `group` setting to the
current process id. This doesn't actually work on Windows, since it
will try to lookup a user whose name is a Fixnum and fail with the
cryptic message:

    no implicit conversion of Fixnum into String

Setting the group for the directory may have been necessary a long time
ago when the test used the real /etc/puppet/ssl directory (when running
as root). However, we changed puppet's spec_helper to always create a
temp directory for each test, so that state from one test doesn't
leak to the next.

And since rspec is the one creating the temp directory, it will always
have permission to read/write.