(PUP-4458) Code clean-up
This commit replaces some strings with constants and alters some conditionals
to make them easier to read. It also changes the Puppet::Parser::Resource
initializer from a method that takes a variable number of arguments (but
really doesn't due to subsequent checks) to a method with three parameters.

The logic remains the same.