(PUP-4752) Improve validation of parameter names
Before this, a badly formed parameter name would go unnoticed. This
causes a problem when not running with --strict_parameters beause
references to a non existing parameter does not cause a failure.
A user may have defined a parameter with 'define foo($Ateam) {}' and
tries to reference it with '$ateam' or '$Ateam' (the former would not
find the value, and the latter would fail with illegal parameter name).

This commit adds validation of the parameter name. It may start with a
lower case letter or underscore, and then any number of word characters.
A new issue code is added, and a new pattern since all existing patterns
(and error messages) are for fully qualified names.