(PUP-4458) Ensure that ensure_in_catalog will accept parameters as array
It's not entirely clear what type the 'parameters' argument has for the
method Puppet::Resource::Type.ensure_in_catalog. The paths found in
Puppet suggests that it is a Hash with key/value pairs but it will also
work if an Array of Puppet::Parser::Resource::Param objects was passed.
This commit ensures that only a Hash will be converted to an Array and
that an Array it will be passed through to the Puppet::Parser::Resource
constructor verbatim.

Please note that if an Array of something that doesn't consist of
Puppet::Parser::Resource::Param objects is passed, that would have failed
in the code preceding this PR as well, since the set_parameter method then
would get nil values due to the each loop in the assign_parameter_values
method.