(PUP-4688) Skip PMT tests on OSX
The system openssl on OSX is integrated with the keystore when
resolving CA certs. However, now that we are compiling and shipping
openssl in AIO, it does not have this capability, so the PMT on OSX
cannot make SSL connections to the forge.  This is a regression
relative to puppet 3.8, and is being tracked in PUP-4822.

For now we skip the test on OSX until the root cause is resolved.