(PUP-4688) Skip user and group acceptance tests on OSX
The user and group providers appear to work on OSX 10.9 and up, but the
acceptance tests fail trying to create users and groups using native
commands, e.g. getent, as part of the tests' preconditions.

This commit skips those tests on OSX until they can be updated as part
of PUP-4824.