(PUP-4688) Define helper methods for stat'ing files
Previously, the acceptance tests executed `stat --format '%U:%G %a`
which doesn't work on osx/bsd. Instead we need to use `stat -f
'%Su:%Sg %p`. The capital `S` ensures we get back names and not
ids. Also the `%p` format specifier returns the raw mode, which
includes the upper file type bit fields, e.g. directory, file,
fifo. That is unlike Linux where `%a` returns only the access rights
in octal.

This commit adds a `stat` helper method that returns a permissions
triplet of owner, group, mode. The latter is returned as an integer
and only contains the access rights, e.g. 0750.

Also, several tests assumed the primary group for the root user was
named "root" though for osx/bsd it's "wheel", etc. This commit adds
`root_user` and `root_group` to account for those difference, and
updates the tests to call it.